+++ Evolve.NET.Core/RoulleteSelection.cs
@@ -5,9 +5,9 @@
         public IChromosome Select(IPopulation population)
         {
             IChromosome chromosome = null;
-            float[] probabilities = GetProbabilities(population);
-            float point = Helper.RandomFloat();
-            float accumulation = 0.0f;
+            double[] probabilities = GetProbabilities(population);
+            double point = Helper.RandomDouble();
+            double accumulation = 0.0f;
             for (int i = 0; i < population.Count; i++)
             {
                 accumulation += probabilities[i];
@@ -21,13 +21,13 @@
             return chromosome;
         }
 
-        private float[] GetProbabilities(IPopulation population)
+        private double[] GetProbabilities(IPopulation population)
         {
-            float sum = 0.0f;
+            double sum = 0.0f;
             for (int i = 0; i < population.Count; i++)
                 sum += population[i].Fitness;
 
-            float[] probabilities = new float[population.Count];
+            double[] probabilities = new double[population.Count];
             for (int i = 0; i < population.Count; i++)
                 probabilities[i] = population[i].Fitness / sum;
